fix wrong trait ID in inherited vtable shims
When emitting an intrinsic of `Virtual` kind (a vtable shim), we
previously looked at the parent item of the trait method to find the
trait ID to include.  This is wrong when calling a method that was
inherited from a supertrait, such as `Display::fmt` on `&dyn Error`: the
vtable shim is for `Display` instances, but the trait object is `Error`,
which (now that traits are fully monomorphized) are unrelated traits
with unrelated vtables.

We now obtain the trait ID from the trait object type, so that
everything matches up properly at run time.