Compute lengths of &str constants properly
`&str` is an unsized type, and the `Layout` for an unsized type has a `size` of
`0`, which isn't what we want. As far as I can tell, the proper way to compute
the size of an unsized type is to use `MPlaceTy`'s `len` function, which sadly
isn't exported. I have copied it here and used it.