saw-rustc: Skip non-functions in init_instances_from_tests without erroring
`crux-rustc` makes it an error to annotate non-functions with `#[crux::test]`
attributes. `saw-rustc` effectivelly annotates _everything_ with that
attribute, but we don't want to error when the annotation is applied to a
non-function, as typical Rust code will have more things (e.g., statics) than
just functions. We now make `saw-rustc` skip non-functions (thereby not marking
them as roots) instead of erroring.

Fixes #55.