Rearrange the way slices are handled.
String slices and ordinary array slices are still special cases, but
they're now handled like references: they contain a def_id that points
to an allocation, which for constants is a static allocation holding
the body. Previously they were handled like values and contained an
inline copy of the actual data.

This makes them come out in Crucible as global references rather than
(magic) constant references; the latter do not support operations that
are needed in practice. See saw-script #2064.

The special cases for bstr are removed; they're no longer needed.

It's probable that at least the body-data special cases for slices are
no longer needed either, but one thing at a time.