Changes in pursuit of the Const_RefRoots problem in crucible-mir.
Very preliminary; expect this commit to disappear before anything gets
merged.

The idea so far is to handle string slices and array slices the same
way as ordinary references, except include the length. For the moment
this means still outputting them with the same "kind" but sharing the
logic for fetching the def_id.

The "bstr" case is also involved for reasons that manifest downstream
in crucible.