rustup: ParamEnv::reveal_all -> empty
@spernsteiner says:

For TypingEnv, I think we can just use TypingEnv::fully_monomorphized(),
since we fully monomorphize everything before we emit it as json.  And
TypingEnv::fully_monomorphized() uses ParamEnv::empty() for the
param_env field, so we can probably use that as a replacement for the
old ParamEnv::reveal_all().  The docs on fully_monomorphized say "Create
a typing environment with no where-clauses in scope where all opaque
types and default associated items are revealed," which sounds like it's
in line with the old reveal_all().

https://github.com/rust-lang/rust/commit/319843d8cd84ee1ec753f836ce3773d44fe0764b