rustup: translate-libs: Key artifacts by src_path
cargo changed the format of package_id in JSON messages in
rust-lang/cargo#13311, but didn't change the format in --unit-graph, so
now they don't match. Use src_path as the key for artifact_outputs
instead. build-script-executed doesn't have src_path, but we can look it
up based on package_id once we have the artifact output.