Properly pretty-print AssertMessages in JSON output
`Assert` statements no longer store strings representing the pretty-printed
message that gets displayed when the assertion fails. Instead, they contain an
`AssertMessage` value, which we must pretty-print to obtain the same string as
before. Alas, the `rustc` API doesn't _quite_ export us the functionality we
need here, so we are forced to reimplement a subset of the `rustc` API
ourselves in pursuit of this.