vtable_descriptor_for_cast: Support vtable casts involving structs
Previously, `vtable_descriptor_for_cast` only supported casts where the from/to
types were references or raw pointers. Besides these two types, it is also
possible to have unsized coercions between structs where exactly one of the
struct's underlying fields differ and require a reference/pointer cast. This is
the sort of cast seen in the example in GaloisInc/crucible#1433.

This adds support for struct-related vtable casts by extending
`vtable_descriptor_for_cast` as needed. When the from/to types are structs, we
recursively walk over the fields of the structs to find a pair of mismatched
field types, which we use to compute the vtable to use in the cast.

Fixes one half of GaloisInc/crucible#1433. (The other half will be on the
`crucible-mir` side.)