vtable_descriptor_for_cast: descend into both ADT fields in tandem
In addition to looking a bit neater, this fixes a bug I recently introduced,
when casting from `&S<S1>` to `&S<dyn Trait>`, where `S` and `S1` were both
structs. We would recursively find both structs' last fields in isolation; for
the target, this would yield `dyn Trait`, as intended, but for the source, this
would yield `S1`'s last field. We would then incorrectly say that the trait
object's concrete type was `S1`'s last field.

We should have found that the trait object's concrete type was `S1` itself. Now,
we do so by descending into both structs synchronously, so that when we descend
to the level of the trait object field of the target, we'll also have descended
to the same level in the source, and will have access to the correct concrete
type to construct the vtable.