crucible: `transmute` pointers in `TypedAllocator` allocation
Instead of using `NonNull::cast` to convert a `NonNull<T>` to a `NonNull<u8>`,
use `core::mem::transmute`, to prevent downstream issues in `crucible-mir`.

`crucible-mir` gives special treatment to several casts, including from `*const
[T; N]` to `*const T`. In Rust, this is a no-op, because Rust makes no
distinction between pointers to arrays and pointers to their first
elements. `crucible-mir`, however, uses an explicitly array-shaped
representation for pointers to arrays, so when it sees such a cast, it needs to
insert an index-0 operation to get the first element of its array and keep its
semantics aligned with Rust's. This is unideal, but often innocuous.

It becomes nocuous when trying to allocate space for `[u8; N]`s - e.g., for a
`Vec<[u8; N]>`. At this type, `TypedAllocator::allocate{,_zeroed}` will create a
`NonNull<[u8; N]>` and cast it to a `NonNull<u8>`. That cast reduces to a
`*const [u8; N]` to `*const u8` cast, which triggers `crucible-mir`'s
aforementioned special casting rule for that case, and causes it to insert a
spurious indexing operation. This makes it impossible to simulate writes of
`[u8; N]`s to that allocation - e.g. via `Vec::push` - since `crucible-mir` has
come to believe it can only hold `u8`s.

A fix for this rightly belongs in `crucible`, but it involves undoing its
special array representation, which is pretty involved.