libs: Simplify optimized implementation of `fmt_u128`
The `Display` impls for `u128` and `i128` are defined in terms of a heavily
optimized `fmt_u128` function that is difficult for `crucible-mir` to
translate. (Among other things, it performs `f64`-to-`u64` conversions and
calls the `write_bytes` intrinsic, neither of which are currently supported.)
We replace this function with a slower (but easier-to-translate),
macro-generated version.

This performs the `mir-json`-side changes that are needed as part of an
eventual fix for https://github.com/GaloisInc/crucible/issues/1509.