Use constant evaluation machinery in ToJson impl for ty::Const
Instead of pattern-matching on the `Leaf` case of a `ValTreeKind`, we instead
leverage the existing constant evaluation machinery for `ConstValue`s. This
approach is simpler and avoids duplicating the logic necessary for writing the
JSON schema for array lengths. (Previously, the `ToJson` impl for `ty::Const`
was only used for array lengths, but with this change, it can now be used for
any sort of constant.)

Preparation for #64.