Remove unsupported initializer from memchr (#194)
* Remove unsupported initializer from memchr

On x86_64, memchr uses an AtomicPtr<()> to
store a casted Function Pointer. This is used
to allow it to do runtime-detection of which
memchr imlementation to use. This change
discards that cleverness and always uses the
fallback. This is ideal because we can't handle
the pointer-cast that appears in the initializer
and which generates translation errors if this
code makes it into the translation.

* Update libs/Patches.md

Co-authored-by: Ryan Scott <rscott@galois.com>

---------

Co-authored-by: Ryan Scott <rscott@galois.com>