Remove CloneShim-related logic for TyKind::Array
As of https://github.com/rust-lang/rust/pull/86041, `Clone` impls for array
types are now implemented in library code rather than using `clone` shims. As
such, the `TyKind::Array` case in the code surrounding `CloneShim`s is now
unreachable. This removes the code to simplify things.

To prevent future regressions, this adds a test case that `clone`s an array and
ensures that it does not emit a `don't know how to build clone shim` warning.

Fixes #197.