Add CloneShim logic for TyKind::FnDef
This avoids producing a `don't know how to build clone shim for FnDef` warning
when compiling code that attempts to `clone` a `FnDef`. As `FnDef`s are
zero-sized, they have no fields that would require additional `clone` methods,
so creating a `clone` shim for a `FnDef` is trivial.

This adds the necessary logic for `TyKind::FnDef` and adds a test case to
ensure that `mir-json` no longer produces a `don't know how to build clone
shim` warning when compiling these sorts of `clone` shims.

Fixes #195.