Reapply "Mark Crucible-related hook functions as #[inline(never)]"
This re-applies the changes from #154, which were omitted from #200 due to an
oversight. It is a bit unclear whether or not `#[inline(never)]` is strictly
necessary here (see
https://github.com/GaloisInc/mir-json/issues/153#issuecomment-3602259898), but
better to be on the safe side.

To make it more likely that we will remember to apply this change in the
future, I have started a "Notes" section in `libs/Patches.md` that very
explicitly calls out the needs for `#[inline(never)]`, and I have cited this
note from the relevant patch descriptions. Going forward, we can use this
"Notes" section to describe other aspects of patches that require more in-depth
explanations that what can be afforded in the brief patch descriptions.