libs/Patches.md: Add an Update convention
In the past, we have forgotten to reapply updates to particular patches (see
https://github.com/GaloisInc/mir-json/issues/153#issuecomment-3602259898 for
one example). This adds a new documentation convention to `libs/Patches.md` to
make this less likely to occur in the future. Whenever we update the
implementation of an existing patch, we now add an *Update* line to the patch
description that summarizes what was updated. Then, when the overall patch is
reapplied during an upgrade to a new Rust toolchain version, we fold the
changes from each *Update* into the main patch, and then we remove the *Update*
line entirely.