libs: Define symbolic_bool primitive function
The previous `Symbolic` impl for `bool` was defined in terms of `symbolic_u8`,
which does not play nicely with `crux-mir-comp`'s override argument matching.
This reimplements the `Symbolic` impl for `bool` in terms of a new
`symbolic_bool` function, which is intended to be overridden in `crucible-mir`.

Towards https://github.com/GaloisInc/saw-script/issues/2872.