libs: Patch MaybeUninit::assume_init() to avoid unsupported pointer cast
https://github.com/rust-lang/rust/commit/8362508989c8941ccbba26aaa8c787cc492c6fef
(which debuted in Rust 1.90.0) changed the implementation of
`MaybeUninit::assume_init()` to use a raw pointer cast for the sake of
improving error messages in Miri. While good for Miri, this is bad for
`crucible-mir`, as this pointer cast can involve distinct Crucible type
representations, which is a type of cast that `crucible-mir` doesn't support.
In particular, if using `array::from_fn()` to construct a length-0 array, then
this would require casting from `UnitRepr` (due to the use of a zero-sized
type) to a `MirAggregateRepr`, which `crucible-mir` does not currently support.

This commit reverts back to the old implementation of `assume_init()`, which
does not have a problematic pointer cast.

Towards https://github.com/GaloisInc/crucible/issues/1674.