from_trait_ref: Fix handling of associated types
Previously, this function would take a subtrait and attempt to compute a
projection for one of its supertraits' associated types by instantiating them
with the type arguments for the subtrait, not the type arguments for the
supertrait. This is utterly wrong, as the type is associated with the
supertrait, not the subtrait. This can cause internal compiler errors when the
two sets of type arguments do not match up, as seen in #211.

This fixes the issue by reworking `from_trait_ref` to better track the type
arguments used to instantiate each supertrait and using them to compute the
projections for each associated type.

Fixes #211.